requirements.txt
appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
click==8.0.1
cryptography==3.4.7
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
flake8-colors==0.1.9
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
identify==2.2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.0
pandas==1.2.5
pathspec==0.8.1
pre-commit==2.13.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
virtualenv==20.4.7
Werkzeug==2.0.1
yapf==0.31.0
